fix(api): fixed method of checking guild manageable
