+++ src/routes/guilds/guild.ts
@@ -1,12 +1,10 @@
 import ApiRequest from '@lib/structures/api/ApiRequest';
 import ApiResponse from '@lib/structures/api/ApiResponse';
+import { canManage } from '@utils/API';
 import { api } from '@utils/Models/Api';
 import { flattenGuild } from '@utils/Models/ApiTransform';
 import { authenticated, ratelimit } from '@utils/util';
-import { Permissions } from 'discord.js';
 import { Route, RouteStore } from 'klasa-dashboard-hooks';
-
-const { FLAGS: { MANAGE_GUILD } } = Permissions;
 
 export default class extends Route {
 
@@ -25,8 +23,7 @@ export default class extends Route {
 		const member = await guild.members.fetch(request.auth!.user_id).catch(() => null);
 		if (!member) return response.error(400);
 
-		const canManage = member.permissions.has(MANAGE_GUILD);
-		if (!canManage) return response.error(401);
+		if (!canManage(guild, member)) return response.error(403);
 
 		const emojis = await api(this.client).guilds(guildID).emojis.get();
 		return response.json({ ...flattenGuild(guild), emojis });
