+++ src/routes/guilds/guild/roles/role.ts
@@ -1,11 +1,9 @@
 import ApiRequest from '@lib/structures/api/ApiRequest';
 import ApiResponse from '@lib/structures/api/ApiResponse';
+import { canManage } from '@utils/API';
 import { flattenRole } from '@utils/Models/ApiTransform';
 import { authenticated, ratelimit } from '@utils/util';
-import { Permissions } from 'discord.js';
 import { Route, RouteStore } from 'klasa-dashboard-hooks';
-
-const { FLAGS: { MANAGE_GUILD } } = Permissions;
 
 export default class extends Route {
 
@@ -24,8 +22,7 @@ export default class extends Route {
 		const member = await guild.members.fetch(request.auth!.user_id).catch(() => null);
 		if (!member) return response.error(400);
 
-		const canManage = member.permissions.has(MANAGE_GUILD);
-		if (!canManage) return response.error(401);
+		if (!canManage(guild, member)) return response.error(403);
 
 		const roleID = request.params.role;
 		const role = guild.roles.get(roleID);
